14/6

+++ src/Controller/peliserieController.js
@@ -37,7 +37,7 @@ router.get('', Authenticate, async (req, res) => {
     return res.status(200).json(peliserie);
 });
 
-router.post('',Authenticate,  async (req, res) => {
+router.post('', Authenticate, async (req, res) => {
     const peliserie = await peliserieService.createPeliserie(req.body);
     const Calificacion = req.body.Calificacion
     if(Calificacion <1 || Calificacion > 5){
@@ -46,29 +46,40 @@ router.post('',Authenticate,  async (req, res) => {
     return res.status(201).json(peliserie);
 });
 
-router.put('/:id',Authenticate,  async (req, res) => {
+router.put('/:id', Authenticate, async (req, res) => {
     const id = req.params.id;
-    const Calificacion = req.body.Calificacion
-
     console.log(id);
-
     if(id != req.body.Id){
         return res.status(400).send();
     }
-    if(Calificacion <1 || Calificacion > 5){
+    const Calificacion = req.body.Calificacion
+    if(Calificacion<1 || Calificacion > 5){
         return res.status(400).send();
     }
-    const peliserie = await peliserieService.updatePeliserieById(Id, );
-    return res.status(200).json(peliserie);
+    //Hacer que si cambiamos solo un atributo se cambie solo eso y no todo JI!
+    const peliserie = await peliserieService.detallePeliserie(id);       
+    
+    const result = await peliserieService.updatePeliserieById(id, fortnite );
+    if(update.rowsAffected[0] == 0){
+        res.status(404).send();
+    }
+    return res.status(200).json(result);
 });
 
-router.delete('/:id',Authenticate,  async (req, res) => {
+router.delete('/:id', Authenticate, async (req, res) => {
     const id = req.params.id;
-    const peliserie = await peliserieService.deletePeliserieById(id);
-    return res.status(200).json(peliserie);
+    if(id < 1){
+        return res.status(400).send();
+    }
+    const result = await peliserieService.deletePeliserieById(id);
+    console.log(result);
+    if(result.rowsAffected[0] == 0){
+        return res.status(404).json(result);
+    }   
+    return res.status(200).json(result);
 });
 
-router.get('/:id',Authenticate,  async (req, res) => {
+router.get('/:id', Authenticate, async (req, res) => {
     const id = req.params.id;
     if(id < 1){
         return res.status(400).send();
